OCUA-17 Extracted the copypasta into a separate class.
Limited the id cells to 3 columns.
Made the popup title more noticeable.
