+++ ui/oce/corruption-risk/style.less
@@ -390,6 +390,9 @@
         hr{
           width: auto;
         }
+        h5{
+          font-weight: bold;
+        }
       }
 
       &:hover {
